+++ internal/handler/handler.go
@@ -1,0 +1,3 @@
+package handler
+
+// Package handler 注册 Gin HTTP 路由处理器