+++ internal/store/store.go
@@ -1,0 +1,3 @@
+package store
+
+// Package store SQLite 数据访问层